+++ x-pack/plugins/search_inference_endpoints/public/components/all_inference_endpoints/tabular_page.tsx
@@ -5,21 +5,15 @@
  * 2.0.
  */
 
-import React, { useCallback, useMemo } from 'react';
-import {} from './constants';
+import React, { useCallback } from 'react';
 
-import { EuiFlexGroup, EuiButton, EuiFlexItem, EuiPageTemplate } from '@elastic/eui';
+import { EuiButton, EuiFlexGroup, EuiFlexItem, EuiPageTemplate } from '@elastic/eui';
 
 import { i18n } from '@kbn/i18n';
 
 import { InferenceAPIConfigResponse } from '@kbn/ml-trained-models-utils';
 
-import {
-  InferenceEndpointUI,
-  INFERENCE_ENDPOINTS_TABLE_PER_PAGE_VALUES,
-  SortFieldInferenceEndpoint,
-  SortOrder,
-} from './types';
+import { useTableData } from './use_table_data';
 
 import { useAllInferenceEndpointsState } from './use_all_inference_endpoints_state';
 
@@ -33,13 +27,6 @@ interface TabularPageProps {
   setIsInferenceFlyoutVisible: (value: boolean) => void;
 }
 
-const getSortField = (field: string): SortFieldInferenceEndpoint => {
-  if (field in SortFieldInferenceEndpoint) {
-    return SortFieldInferenceEndpoint[field as keyof typeof SortFieldInferenceEndpoint];
-  }
-  return SortFieldInferenceEndpoint.endpoint;
-};
-
 export const TabularPage: React.FC<TabularPageProps> = ({
   addEndpointLabel,
   setIsInferenceFlyoutVisible,
@@ -47,76 +34,28 @@ export const TabularPage: React.FC<TabularPageProps> = ({
 }) => {
   const { queryParams, setQueryParams } = useAllInferenceEndpointsState();
 
-  const sorting = useMemo(
-    () => ({
-      sort: {
-        direction: queryParams.sortOrder,
-        field: queryParams.sortField,
-      },
-    }),
-    [queryParams.sortField, queryParams.sortOrder]
+  const { paginatedSortedTableData, pagination, sorting } = useTableData(
+    inferenceEndpoints,
+    queryParams
   );
 
-  const pagination = useMemo(
-    () => ({
-      pageIndex: queryParams.page - 1,
-      pageSize: queryParams.perPage,
-      pageSizeOptions: INFERENCE_ENDPOINTS_TABLE_PER_PAGE_VALUES,
-      totalItemCount: inferenceEndpoints.length ?? 0,
-    }),
-    [inferenceEndpoints, queryParams]
-  );
-
-  const tableOnChangeCallback = useCallback(
+  const handleTableChange = useCallback(
     ({ page, sort }) => {
-      let newQueryParams = queryParams;
-      if (sort) {
-        newQueryParams = {
-          ...newQueryParams,
-          sortField: getSortField(sort.field),
+      const newQueryParams = {
+        ...queryParams,
+        ...(sort && {
+          sortField: sort.field,
           sortOrder: sort.direction,
-        };
-      }
-      if (page) {
-        newQueryParams = {
-          ...newQueryParams,
+        }),
+        ...(page && {
           page: page.index + 1,
           perPage: page.size,
-        };
-      }
+        }),
+      };
       setQueryParams(newQueryParams);
     },
     [queryParams, setQueryParams]
   );
-
-  const tableData: InferenceEndpointUI[] = useMemo(() => {
-    return inferenceEndpoints.map((endpoint) => ({
-      endpoint: endpoint.model_id,
-      provider: endpoint.service,
-      type: endpoint.task_type,
-    }));
-  }, [inferenceEndpoints]);
-
-  const sortedTableData: InferenceEndpointUI[] = useMemo(() => {
-    return [...tableData].sort((a, b) => {
-      const aValue = a[queryParams.sortField];
-      const bValue = b[queryParams.sortField];
-
-      if (aValue < bValue) {
-        return queryParams.sortOrder === SortOrder.desc ? 1 : -1;
-      }
-      if (aValue > bValue) {
-        return queryParams.sortOrder === SortOrder.asc ? -1 : 1;
-      }
-      return 0;
-    });
-  }, [tableData, queryParams]);
-
-  const paginatedSortedTableData: InferenceEndpointUI[] = useMemo(() => {
-    const startIndex = pagination.pageIndex * pagination.pageSize;
-    const endIndex = startIndex + pagination.pageSize;
-    return sortedTableData.slice(startIndex, endIndex);
-  }, [sortedTableData, pagination]);
 
   return (
     <InferenceEndpointsProvider>
@@ -157,7 +96,7 @@ export const TabularPage: React.FC<TabularPageProps> = ({
         <EndpointsTable
           columns={TABLE_COLUMNS}
           data={paginatedSortedTableData}
-          onChange={tableOnChangeCallback}
+          onChange={handleTableChange}
           pagination={pagination}
           sorting={sorting}
         />
